Update Adafruit library for SGP30 sensor
