--- lib/Adafruit_SGP30-1.0.0.13/.gitignore
@@ -1,4 +1,0 @@
-*~
-Doxyfile*
-doxygen_sqlite3.db
-html
