+++ lib/Adafruit_SGP30-1.0.5/.gitignore
@@ -1,0 +1,11 @@
+*~
+Doxyfile*
+doxygen_sqlite3.db
+html# osx
+.DS_Store
+
+# doxygen
+Doxyfile*
+doxygen_sqlite3.db
+html
+*.tmp
