+++ lib/Adafruit_SGP30-1.0.5/.travis.yml
