+++ lib/Adafruit_SGP30-1.0.5/library.properties
@@ -1,5 +1,5 @@
 name=Adafruit SGP30 Sensor
-version=1.0.2
+version=1.0.5
 author=Adafruit
 maintainer=Adafruit <info@adafruit.com>
 sentence=This is an Arduino library for the Adafruit SGP30 Gas / Air Quality Sensor
