+++ lib/Adafruit_SGP30-1.0.5/license.txt
